storage interface
Keep the original storage interface and implement plugin related logic
in the plugin's main packages.
