+++ storage/s3/s3.go
@@ -12,8 +12,6 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/minio/minio-go/v7"
-	"github.com/minio/minio-go/v7/pkg/credentials"
-	"github.com/mitchellh/mapstructure"
 
 	"github.com/connylabs/ingest"
 	"github.com/connylabs/ingest/storage"
@@ -36,52 +34,15 @@ type minioStorage struct {
 }
 
 // New returns a new Storage that can store objects to S3.
-func New() storage.Storage {
-	return &minioStorage{}
-}
-
-const defaultEndpoint = "s3.amazonaws.com"
-
-type sourceConfig struct {
-	Endpoint        string
-	Insecure        bool
-	AccessKeyID     string `json:"accessKeyID"`
-	SecretAccessKey string
-	Bucket          string
-	Prefix          string
-	Recursive       bool
-}
-
-type destinationConfig struct {
-	sourceConfig    `mapstructure:",squash"`
-	MetafilesPrefix string
-}
-
-func (ms *minioStorage) Configure(config map[string]interface{}) error {
-	dc := new(destinationConfig)
-	err := mapstructure.Decode(config, dc)
-	if err != nil {
-		return nil
+func New(bucket, prefix, metafilesPrefix string, mc MinioClient, useDone bool, l log.Logger) storage.Storage {
+	return &minioStorage{
+		bucket:          bucket,
+		mc:              mc,
+		l:               l,
+		prefix:          prefix,
+		metafilesPrefix: metafilesPrefix,
+		useDone:         useDone,
 	}
-	if dc.Endpoint == "" {
-		dc.Endpoint = defaultEndpoint
-	}
-	mc, err := minio.New(dc.Endpoint, &minio.Options{
-		Creds:  credentials.NewStaticV4(dc.AccessKeyID, dc.SecretAccessKey, ""),
-		Secure: !dc.Insecure,
-	})
-	if err != nil {
-		return err
-	}
-
-	ms.bucket = dc.Bucket
-	ms.mc = mc
-	ms.l = log.NewNopLogger()
-	ms.prefix = dc.Prefix
-	ms.metafilesPrefix = dc.MetafilesPrefix
-	ms.useDone = dc.MetafilesPrefix != ""
-
-	return nil
 }
 
 func (ms *minioStorage) Stat(ctx context.Context, element ingest.Codec) (*storage.ObjectInfo, error) {
